requirements.txt
aiohttp==3.4.4
appdirs==1.4.3
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
factory-boy==2.11.1
Faker==1.0.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
multidict==4.5.2
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pytest==4.0.1
pytest-factoryboy==2.0.2
pytest-flask==0.14.0
pytest-mock==1.10.0
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.14
text-unidecode==1.2
toml==0.10.0
urllib3==1.24.1
Werkzeug==0.14.1
yarl==1.2.6
